+++ portainer-monitor-addon/CHANGELOG.md
@@ -1,6 +1,7 @@
 ﻿## Changelog 🚀
 ### 1.2.0
 - Add support for direct `Portainer Agent` connection
+- Limit Portainer API `system/version` requests for update checks to prevent `GitHub API rate limit exceeded`, since on each request Portainer will get the latest version info via GitHub API
 
 ### 1.1.1
 - Improved perfomance for resource monitoring